requirements.txt
nltk
numpy
tensorflow-cpu
flask
setuptools
